JoinMailingList: moved to shared components
- exports NewsLetterPledge to be shared
- used NewsLetterPledge in MailingListToggle
