+++ src/components/index.js
@@ -3,9 +3,11 @@ export * from "./NavBar";
 export * from "./FormTools";
 export * from "./CourseCard";
 export * from "./IconMessage";
+export * from "./JoinMailingList";
 
 export { default as Logo } from "./Logo";
 export { default as NavBar } from "./NavBar";
 export { default as CourseCard } from "./CourseCard";
 export { default as IconMessage } from "./IconMessage";
 export { default as QueryHandler } from "./QueryHandler";
+export { default as JoinMailingList } from "./JoinMailingList";
